Rearrange LensKit CLI into org.lenskit.cli package
