+++ lenskit-cli/build.gradle
@@ -64,7 +64,7 @@ meta {
 // Build an application
 apply plugin: 'application'
 
-mainClassName = 'org.grouplens.lenskit.cli.Main'
+mainClassName = 'org.lenskit.cli.Main'
 applicationName = 'lenskit'
 
 // Render the manpages
