+++ lenskit-cli/src/main/java/org/lenskit/cli/commands/Graph.java
@@ -18,7 +18,7 @@
  * this program; if not, write to the Free Software Foundation, Inc., 51
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package org.grouplens.lenskit.cli;
+package org.lenskit.cli.commands;
 
 import com.google.auto.service.AutoService;
 import net.sourceforge.argparse4j.inf.ArgumentParser;
@@ -33,6 +33,8 @@ import org.grouplens.lenskit.data.dao.EventDAO;
 import org.grouplens.lenskit.data.pref.PreferenceDomain;
 import org.grouplens.lenskit.eval.graph.GraphDumper;
 import org.grouplens.lenskit.inject.RecommenderGraphBuilder;
+import org.lenskit.cli.Command;
+import org.lenskit.cli.util.ScriptEnvironment;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
