+++ lenskit-cli/src/main/java/org/lenskit/cli/util/Logging.java
@@ -18,7 +18,7 @@
  * this program; if not, write to the Free Software Foundation, Inc., 51
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package org.grouplens.lenskit.cli;
+package org.lenskit.cli.util;
 
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
